+++ docs/changelog.rst
@@ -6,6 +6,11 @@ Changelog
 `Unreleased`_
 -------------
 
+Fixed
+~~~~~
+
+- Handling ``nullable`` attribute in Open API schemas. `#335`_
+
 `0.20.3`_ - 2019-12-17
 ----------------------
 
@@ -526,6 +531,7 @@ Fixed
 .. _0.3.0: https://github.com/kiwicom/schemathesis/compare/v0.2.0...v0.3.0
 .. _0.2.0: https://github.com/kiwicom/schemathesis/compare/v0.1.0...v0.2.0
 
+.. _#335: https://github.com/kiwicom/schemathesis/issues/335
 .. _#332: https://github.com/kiwicom/schemathesis/issues/332
 .. _#330: https://github.com/kiwicom/schemathesis/issues/330
 .. _#322: https://github.com/kiwicom/schemathesis/issues/322